refactor error handling to separate file
