+++ src/parser/parser.c
@@ -1,73 +1,10 @@
 #include "cub3D.h"
 
-// Read-Errors
-#define OPT_UNKNWON "Found unexpected option"
-#define NO_MAP "File doesn't contain valid map"
-#define NO_OPTS "File has no options"
-#define MAP_NOT_LAST "No options specified before map"
-#define MULTIPLE_MAPS "File contains multiple maps"
-// Parse-Errors
-#define INVALID_TEXTURE "File contains invalid texture"
-#define INVALID_COLOR "Options contain invalid color"
-#define TOO_MANY_TEXTURES "File contains too many textures"
-#define TOO_MANY_SPAWNS "File contains too many spawn positions"
-#define TOO_MANY_COLORS "File contains too many color options"
-// Validation-Errors
-#define MISSING_OPTION "File misses a mandatory option"
-#define INVALID_MAP "Map is invalid - not playable"
-
 static t_map			*init_map(void);
 static t_file_content	*read_file(int fd, t_map *map,
 							t_file_content *file_content);
 static void				populate_map(t_list *line_list, t_map *map);
 static t_cube_type		**create_map(t_list *line_list, t_map *map);
-
-void	print_error(enum e_parser_error err)
-{
-	ft_putendl_fd("Error", STDERR_FILENO);
-	if (err == opt_unknown)
-		ft_putendl_fd(OPT_UNKNWON, STDERR_FILENO);
-	else if (err == no_map)
-		ft_putendl_fd(NO_MAP, STDERR_FILENO);
-	else if (err == multiple_maps)
-		ft_putendl_fd(MULTIPLE_MAPS, STDERR_FILENO);
-	else if (err == map_not_last)
-		ft_putendl_fd(MAP_NOT_LAST, STDERR_FILENO);
-	else if (err == no_opts)
-		ft_putendl_fd(NO_OPTS, STDERR_FILENO);
-	else if (err == invalid_texture)
-		ft_putendl_fd(INVALID_TEXTURE, STDERR_FILENO);
-	else if (err == invalid_color)
-		ft_putendl_fd(INVALID_COLOR, STDERR_FILENO);
-	else if (err == too_many_textures)
-		ft_putendl_fd(TOO_MANY_TEXTURES, STDERR_FILENO);
-	else if (err == too_many_spawns)
-		ft_putendl_fd(TOO_MANY_SPAWNS, STDERR_FILENO);
-	else if (err == too_many_colors)
-		ft_putendl_fd(TOO_MANY_COLORS, STDERR_FILENO);
-	else if (err == missing_option)
-		ft_putendl_fd(MISSING_OPTION, STDERR_FILENO);
-	else if (err == invalid_map)
-		ft_putendl_fd(INVALID_MAP, STDERR_FILENO);
-}
-
-void	parser_error(t_file_content *file_content, t_map *map)
-{
-	print_error(map->state->error_type);
-	if (file_content)
-		free_filecontent(file_content);
-	if (map->state->map_parsed && map->cubes && *(map)->cubes)
-		free_cubes(map);
-	if (map)
-	{
-		if (map->start_dir)
-			free(map->start_dir);
-		if (map->start_pos)
-			free(map->start_pos);
-		free(map->state);
-		free(map);
-	}
-}
 
 // TODO: needs proper freeing on errors
 t_map	*parse(int fd)
